ckb-auth-cli changed to lib * main.rs move to bin/tools.rs * the original code that supports the chain is used as the interface of the lib
